+++ src/fs/azure.ts
@@ -27,6 +27,7 @@ import * as vscode from 'vscode';
 import * as vscode_helpers from 'vscode-helpers';
 import * as vscrw from '../extension';
 import * as vscrw_fs from '../fs';
+import * as vscrw_search from '../search';
 
 interface AzureBlobConnection {
     account: string;
@@ -407,10 +408,15 @@ export class AzureBlobFileSystem extends vscrw_fs.FileSystemBase {
      * @param {vscode.ExtensionContext} context The extension context.
      */
     public static register(context: vscode.ExtensionContext) {
+        const FS_PROVIDER = new AzureBlobFileSystem();
+
         context.subscriptions.push(
             vscode.workspace.registerFileSystemProvider(AzureBlobFileSystem.scheme,
-                                                        new AzureBlobFileSystem(),
-                                                        { isCaseSensitive: true })
+                                                        FS_PROVIDER,
+                                                        { isCaseSensitive: true }),
+
+            vscode.workspace.registerSearchProvider(AzureBlobFileSystem.scheme,
+                                                    new vscrw_search.FileSystemSearchProvider(FS_PROVIDER)),
         );
     }