+++ src/fs/dropbox.ts
@@ -23,6 +23,7 @@ import * as vscode from 'vscode';
 import * as vscode_helpers from 'vscode-helpers';
 import * as vscrw from '../extension';
 import * as vscrw_fs from '../fs';
+import * as vscrw_search from '../search';
 
 interface DropboxConnection {
     client: Dropbox.Dropbox;
@@ -257,10 +258,15 @@ export class DropboxFileSystem extends vscrw_fs.FileSystemBase {
      * @param {vscode.ExtensionContext} context The extension context.
      */
     public static register(context: vscode.ExtensionContext) {
+        const FS_PROVIDER = new DropboxFileSystem();
+
         context.subscriptions.push(
             vscode.workspace.registerFileSystemProvider(DropboxFileSystem.scheme,
-                                                        new DropboxFileSystem(),
-                                                        { isCaseSensitive: false })
+                                                        FS_PROVIDER,
+                                                        { isCaseSensitive: false }),
+
+            vscode.workspace.registerSearchProvider(DropboxFileSystem.scheme,
+                                                    new vscrw_search.FileSystemSearchProvider(FS_PROVIDER)),
         );
     }