+++ src/fs/slack.ts
@@ -25,6 +25,7 @@ import * as vscode from 'vscode';
 import * as vscode_helpers from 'vscode-helpers';
 import * as vscrw from '../extension';
 import * as vscrw_fs from '../fs';
+import * as vscrw_search from '../search';
 
 interface SlackConnection {
     channel: string;
@@ -275,10 +276,15 @@ export class SlackFileSystem extends vscrw_fs.FileSystemBase {
      * @param {vscode.ExtensionContext} context The extension context.
      */
     public static register(context: vscode.ExtensionContext) {
+        const FS_PROVIDER = new SlackFileSystem();
+
         context.subscriptions.push(
             vscode.workspace.registerFileSystemProvider(SlackFileSystem.scheme,
-                                                        new SlackFileSystem(),
-                                                        { isCaseSensitive: false })
+                                                        FS_PROVIDER,
+                                                        { isCaseSensitive: false }),
+
+            vscode.workspace.registerSearchProvider(SlackFileSystem.scheme,
+                                                    new vscrw_search.FileSystemSearchProvider(FS_PROVIDER)),
         );
     }